% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_to_ombudsman.R
\name{translate_to_ombudsman}
\alias{translate_to_ombudsman}
\title{Translates text to Canes Ombudsman-speak}
\usage{
translate_to_ombudsman(text)
}
\arguments{
\item{text}{A string of text to be translated to Ombudsman. Text must be within quotes (" ").}
}
\description{
Returns text translated to Canes Ombudsman-speak
}
\examples{

translate_to_ombudsman("I would love to forgive John Forslund")

translate_to_ombudsman("Would you consider trading William Nylander for Sebastian Aho 1 for 1?")
  
}
